add a version with progress bar
Just for palying around with a progress bar.

It turned out, progress bars not that easy if there is additional
output. This makes it necessary to go up and down the lines. Before
making a change to the bar, go up to the position of the bar,
afterwards go down again.

Except for dropping color support this does the same as the two other
verions. Colour could be supported with the console crate, so maybe it
would be a good thing to add real color, that is even save to use (at
least a litte bit).